update prompt name column, fix data type for repo, and add a response to the main page

+++ services/prompt-service/src/main/java/com/psognathus/code/prompt_service/prompt/PromptRepository.java
@@ -3,5 +3,5 @@ package com.psognathus.code.prompt_service.prompt;
 import com.psognathus.code.prompt_service.prompt.Prompt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromptRepository extends JpaRepository<Prompt, Integer> {
+public interface PromptRepository extends JpaRepository<Prompt, Long> {
 }
